backend/requirements.txt
fastapi
uvicorn
pymongo[srv]
python-dotenv
requests
together
python-multipart
bson
reportlab